fix: correct type for models

+++ app/domain/post/post.go
@@ -10,7 +10,7 @@ type Post struct {
 	ID        int64  `json:"id"`
 	Title     string `json:"title"`
 	Content   string `json:"content"`
-	AuthorID  int
+	AuthorID  int64
 	Author    author.Author `json:"author"`
 	CreatedAt time.Time     `json:"created_at"`
 	UpdatedAt time.Time     `json:"updated_at"`
